Create Controller superclass & move to lib.control
